Refactor: Restructure domain interfaces and base classes
Refactored and reorganized interfaces and base classes to improve modularity and align with domain-driven design principles.

- Moved `IEntity` to `PHMIS.Domain.Common.Interfaces` and added an `IsPublic` property.
- Relocated `IAuditableCreatedDateAndCreatedBy` and `IAuditableEntity` to `PHMIS.Domain.Common.Interfaces`.
- Moved `BaseEntity` to `PHMIS.Domain.Common.BaseAbstract` without structural changes.
- Redefined `BaseAuditableEntity` with new properties (`CreatedBy`, `CreatedDate`, `UpdatedBy`, `UpdatedDate`) and default values.
- Added `AuditableCreatedDateAndCreatedBy` class implementing `IAuditableCreatedDateAndCreatedBy`.
- Introduced `BaseAuditableForJoinedEntity` class implementing `IAuditableEntity` without an `Id` property.
- Updated `Hospital`, `LabTest`, `LabTestGroup`, `Patient`, and `Province` to use the `BaseEntity` from `PHMIS.Domain.Common.BaseAbstract`.
- Added a placeholder `Class1` in `PHMIS.Domain.Common.Interfaces`.

These changes aim to improve code organization, modularity, and maintainability.
