--- PHMIS.Application/Common/Interfaces/IEntity.cs
@@ -1,11 +1,0 @@
-﻿
-namespace PHMIS.Application.Common.Interfaces
-
-{
-    public interface IEntity
-    {
-        public int Id { get; set; }
-        public string? Name { get; set; }
-
-    }
-}
