+++ PHMIS.Domain/Common/BaseAbstract/BaseAuditableForJoinedEntity.cs
@@ -1,0 +1,14 @@
+﻿using PHMIS.Domain.Common.BaseAbstract;
+using PHMIS.Domain.Common.Interfaces;
+
+namespace PHMIS.Domain.Common.BaseAbstract
+{
+    public class BaseAuditableForJoinedEntity : IAuditableEntity
+    {
+        //public int Id { get; set; }
+        public string? CreatedBy { get; set; }
+        public DateTime? CreatedDate { get; set; }
+        public int? UpdatedBy { get; set; }
+        public DateTime? UpdatedDate { get; set; }
+    }
+}
