+++ PHMIS.Domain/Common/Interfaces/Class1.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PHMIS.Domain.Common.Interfaces
+{
+    internal class Class1
+    {
+    }
+}
