+++ PHMIS.Domain/Common/Interfaces/IAuditableEntity.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PHMIS.Application.Common.Interfaces;
+namespace PHMIS.Domain.Common.Interfaces;
 
 
 public interface IAuditableEntity /*: IEntity*/
