+++ PHMIS.Domain/Common/Interfaces/IEntity.cs
@@ -1,0 +1,12 @@
+﻿
+namespace PHMIS.Domain.Common.Interfaces;
+
+
+public interface IEntity
+{
+    public int Id { get; set; }
+    public bool IsPublic { get; set; }
+    public string? Name { get; set; }
+
+}
+
